requirements.txt

Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27.0
drf_yasg
django-cors-header==4.7.0